Added support for variables in assembler
You can now declare variables using the [value:name] syntax, which
simply take up one byte. (actually, if you declare a variable higher
than 255, the size is increased to the smallest possible number of
bytes to represent that number. Names are optional. ) As well as
reference their address using the .[name] syntax. An example is
available in the Main file.